requirements.txt
# SI²A: Shadow IT Incident Autopilot - Dependencies
# BigQuery AI Hackathon Project

# Core BigQuery and BigFrames
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-dataframes>=1.0.0

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Google Cloud SDK
google-cloud-storage>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0

# Additional utilities
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
